Also add a new non-regression test to avoid the problem to come back again.

+++ src/common/ip.h
@@ -391,8 +391,8 @@ uint32_t ipv6_get_flow_label(const struct ip_packet *ip);
 const struct in6_addr * ipv6_get_saddr(const struct ip_packet *ip);
 const struct in6_addr * ipv6_get_daddr(const struct ip_packet *ip);
 void ipv6_set_flow_label(struct ip_packet *const ip, const uint32_t value);
-uint8_t ip_get_extension_size(const unsigned char *ext);
-uint8_t ip_get_total_extension_size(const struct ip_packet *ip);
+unsigned short ip_get_extension_size(const unsigned char *ext);
+unsigned short ip_get_total_extension_size(const struct ip_packet *ip);
 
 /* Private functions (do not use directly) */
 int get_ip_version(const unsigned char *packet,
